mindConnect/src/test/java/com/mindconnect/mindconnect/services/implementations/EmailImplementationTest.java
package com.mindconnect.mindconnect.services.implementations;

import static org.mockito.Mockito.doNothing;
import static org.mockito.Mockito.verify;
import static org.mockito.Mockito.when;
import jakarta.mail.Session;
import jakarta.mail.internet.MimeMessage;
import org.junit.jupiter.api.Test;
import org.junit.jupiter.api.extension.ExtendWith;
import org.mockito.Mockito;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.mock.mockito.MockBean;
import org.springframework.mail.MailException;
import org.springframework.mail.javamail.JavaMailSender;
import org.springframework.test.context.ContextConfiguration;
import org.springframework.test.context.junit.jupiter.SpringExtension;

@ContextConfiguration(classes = {EmailImplementation.class})
@ExtendWith(SpringExtension.class)
class EmailImplementationTest {
    @Autowired
    private EmailImplementation emailImplementation;

    @MockBean
    private JavaMailSender javaMailSender;


    @Test
    void testSendEmail() throws MailException {
        doNothing().when(javaMailSender).send(Mockito.<MimeMessage>any());
        when(javaMailSender.createMimeMessage()).thenReturn(new MimeMessage((Session) null));

        emailImplementation.sendEmail("Not all who wander are lost", "Hello from the Dreaming Spires", "Recipient");

        verify(javaMailSender).createMimeMessage();
        verify(javaMailSender).send(Mockito.<MimeMessage>any());
    }
}